Fix import errors and organize code better
